divide SpotifyApiClient into impl and client
